fuel_upgrade: add apply puppet fixes hack
The fuel_upgrade script should be able to deliver fixes for puppet
manifests, since it's critical for OpenStack patching feature. The
system should be flexible and adding new puppet fixes should make
no pain.

Since now, the script searches a special folders in tarball for fixes
and install them if they are found.

Change-Id: Ic5f6881cab8c047b55dc0edec8ce1e377c776b17
Blueprint: upgrade-to-5-1