fuel_upgrade: patch 'dhcrelay_monitor' wrapper
Since Fuel 6.1 we're using docker with host networking, so we don't
require 'dhcrelay' anymore. Still, if something goes wrong and rollback
was performed, we need to launch dhcrelay again. In order to do it
properly, we need to get cobbler container's ip address, but
we don't have hooks in 'dockerctl' anymore. So we have to inject
this hook directly into 'dhcrelay_monitor' wrapper.

Change-Id: I7ee2f65e584c4c2dae757a4a301fa34879656400
Closes-Bug: #1422787