Backup all data before upgrade and stop supervisor
Unfortunately, we have a puppet run inside containers on each start and
that's cause to the situation when old containers install new packages
on restart. We don't have this issue before because our containers won't
restarted during upgrade process and now they do (because of upgrading
Docker package).

As workaround, we have to:

* backup all data before starting upgrades
* stop supervisor before host system puppet run, so containers won't be
  restarted and hence won't install new packages

Still, it's hackish.. but I don't know better solution now :(

Closes-Bug: #1455419

Change-Id: I9522ab561718473873faa0f38c4cb34c3bee64b9
Signed-off-by: Igor Kalnitsky <igor@kalnitsky.org>